+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Allow changing field in import and export reports ([#244](../../pull/244))
 - Allow redacting and approving lists of images ([#247](../../pull/247))
 - Allow schemas that do not require file names ([#250](../../pull/250))
+- Allow changing fields used for image and folder names ([#261](../../pull/261))
 
 ### Improvements
 - Added more logging during export ([#222](../../pull/222))